requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
rich==13.9.4
shellingham==1.5.4
sniffio==1.3.1
starlette==0.41.3
typer==0.13.1
typing_extensions==4.12.2
uvicorn==0.32.1
watchfiles==0.24.0
websockets==14.1
Jinja2==3.1.3
Pillow==10.0.0 
tiktoken==0.6.0
pydantic-settings==2.6.0
sqlmodel==0.0.22
passlib==1.7.4
emails==0.6
pyjwt==2.10.1
cryptography==44.0.1
azure-communication-email==1.0.0
alembic==1.14.0
bcrypt==4.0.1
tenacity==9.0.0
asyncpg==0.30.0
sqladmin[full]==0.20.1
gunicorn==23.0.0
celery==5.4.0
azure-identity==1.19.0
azure-servicebus==7.13.0
fastapi-sso==0.18.0